+++ src/kudu/tserver/scanners.cc
@@ -119,7 +119,7 @@ bool ScannerManager::UnregisterScanner(const string& scanner_id) {
   return scanners_by_id_.erase(scanner_id) > 0;
 }
 
-size_t ScannerManager::CountActiveScanners() const {
+uint64_t ScannerManager::CountActiveScanners() const {
   boost::shared_lock<boost::shared_mutex> l(lock_);
   return scanners_by_id_.size();
 }
